requirements.txt
faker
pandas
python-dotenv
sqlalchemy
tqdm
numpy